pyproject.toml
[project]
name = "rng-viz"
version = "0.1.0"
description = "TrueRNG Pro V2 bitstream analyzer and visualizer inspired by the Global Consciousness Project"
authors = [
    {name = "Chase", email = ""}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "textual>=0.50.0",
    "rich>=13.0.0",
    "pyserial>=3.5",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "click>=8.1.0",
    "pandas>=2.1.0",
    "asyncio-mqtt>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
rng-viz = "rng_viz.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
